Square Drive now working
Adjusted Gyro some, added encoder to drive straight for distance, still
could be more repeatable. turn on carpet with non-PID control seems to
not be repeatable.  depends on Gyro Gain for sure and maybe other
factors. moved Gyro to middle of the robot, might have helped or at
least changed the gyro angle/volts measurement.  This is still copied in
with older code not sure it if would compile