Motion Control code is working a little still needs debugging
Implemented code in Turn 90 Degrees to test the Motion Control code, it
still needs more work but it does do something, it turns the robot and
at least now it is not oscelating widely like happened on the first
attempt. but it will start turning wrong way and just go on forever.
otherwise it just ends up maxing the power and the rotation rate is not
hit (might have been that the battery was getting low.